#10 カレンダー修正しました。その他調整・追加しました。

+++ billionaire/src/main/java/web/fixedcost/FixedCostDeleateController.java
@@ -6,6 +6,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.SessionAttributes;
+import org.springframework.web.context.request.WebRequest;
 
 import service.fixedcost.FixedCostDeleateService;
 
@@ -17,7 +18,11 @@ public class FixedCostDeleateController {
 	private FixedCostDeleateService fixedCostDeleateService;
 
 	@RequestMapping(method=RequestMethod.POST, params="fixedCostId")
-	public String delete(@RequestParam("fixedCostId") Integer fixedCostId){
+	public String delete(@RequestParam("fixedCostId") Integer fixedCostId, WebRequest webRequest ){
+		if(webRequest.getAttribute("user", WebRequest.SCOPE_SESSION) == null){
+			return "redirect:../login";
+		}
+
 		fixedCostDeleateService.delete(fixedCostId);
 		return "redirect:../register/fixed_cost/list";
 	}
